core/libs/bridges/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bridges"
version = "1.0.0"
description = "CoratiaOS serial bridge helpers"
requires-python = ">=3.11"
dependencies = ["pyserial"]

[tool.setuptools.packages.find]
where = ["src"]
